+++ telers/src/types/unique_gift_colors.rs
@@ -11,11 +11,11 @@ pub struct UniqueGiftColors {
     /// Main color used in light themes; RGB format
     pub light_theme_main_color: i64,
     /// List of 1-3 additional colors used in light themes; RGB format
-    pub light_theme_other_colors: Box<[u8]>,
+    pub light_theme_other_colors: Box<[i64]>,
     /// Main color used in dark themes; RGB format
     pub dark_theme_main_color: i64,
     /// List of 1-3 additional colors used in dark themes; RGB format
-    pub dark_theme_other_colors: Box<[u8]>,
+    pub dark_theme_other_colors: Box<[i64]>,
 }
 impl UniqueGiftColors {
     /// Creates a new `UniqueGiftColors`.
@@ -32,10 +32,10 @@ impl UniqueGiftColors {
         T0: Into<Box<str>>,
         T1: Into<Box<str>>,
         T2: Into<i64>,
-        T3Item: Into<u8>,
+        T3Item: Into<i64>,
         T3: IntoIterator<Item = T3Item>,
         T4: Into<i64>,
-        T5Item: Into<u8>,
+        T5Item: Into<i64>,
         T5: IntoIterator<Item = T5Item>,
     >(
         model_custom_emoji_id: T0,
@@ -87,7 +87,7 @@ impl UniqueGiftColors {
     /// # Notes
     /// Adds multiple elements.
     #[must_use]
-    pub fn light_theme_other_colors<T: Into<Box<[u8]>>>(mut self, val: T) -> Self {
+    pub fn light_theme_other_colors<T: Into<Box<[i64]>>>(mut self, val: T) -> Self {
         self.light_theme_other_colors = self
             .light_theme_other_colors
             .into_vec()
@@ -102,7 +102,7 @@ impl UniqueGiftColors {
     /// # Notes
     /// Adds a single element.
     #[must_use]
-    pub fn light_theme_other_color<T: Into<u8>>(mut self, val: T) -> Self {
+    pub fn light_theme_other_color<T: Into<i64>>(mut self, val: T) -> Self {
         self.light_theme_other_colors = self
             .light_theme_other_colors
             .into_vec()
@@ -124,7 +124,7 @@ impl UniqueGiftColors {
     /// # Notes
     /// Adds multiple elements.
     #[must_use]
-    pub fn dark_theme_other_colors<T: Into<Box<[u8]>>>(mut self, val: T) -> Self {
+    pub fn dark_theme_other_colors<T: Into<Box<[i64]>>>(mut self, val: T) -> Self {
         self.dark_theme_other_colors = self
             .dark_theme_other_colors
             .into_vec()
@@ -139,7 +139,7 @@ impl UniqueGiftColors {
     /// # Notes
     /// Adds a single element.
     #[must_use]
-    pub fn dark_theme_other_color<T: Into<u8>>(mut self, val: T) -> Self {
+    pub fn dark_theme_other_color<T: Into<i64>>(mut self, val: T) -> Self {
         self.dark_theme_other_colors = self
             .dark_theme_other_colors
             .into_vec()
